Move park and ride popup content to sidebar
