+++ app/util/path.js
@@ -18,6 +18,7 @@ export const PREFIX_TIMETABLE = 'fahrplan';
 export const PREFIX_ROADWORKS = 'baustellen';
 export const PREFIX_BIKE_PARKS = 'fahrradparkplaetze';
 export const PREFIX_CHARGING_STATIONS = 'ladestationen';
+export const PREFIX_PARK_AND_RIDE_FACILITIES = 'parkandride';
 export const stopUrl = id => id;
 export const LOCAL_STORAGE_EMITTER_PATH = '/local-storage-emitter';
